Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/accessing-other-browsing-contexts/indexed-browsing-contexts-03-expected.txt


FAIL Indexed child browsing contexts assert_equals: The first child browsing context's container should be the object element. expected null but got object "[object Window]"
